enable maximal warnings (and warnings as errors) on Windows
Manually exclude each warning emitted by the HM code that we chose to ignore,
fix one warning to avoid ignoring all warnings of that type in the App folder.
Explicitly disable some dangerous warnings directly in the files which emit
them, so they can remain enabled globally to catch bugs as they are created.

+++ source/App/TAppDecoder/TAppDecTop.cpp
@@ -374,7 +374,7 @@ Bool TAppDecTop::isNaluWithinTargetDecLayerIdSet( InputNALUnit* nalu )
   }
   for (std::vector<Int>::iterator it = m_targetDecLayerIdSet.begin(); it != m_targetDecLayerIdSet.end(); it++)
   {
-    if ( nalu->m_reservedZero6Bits == (*it) )
+    if ( nalu->m_reservedZero6Bits == (UInt) (*it) )
     {
       return true;
     }
